Move AsyncManualResetEvent to Tools
